pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "basler-camera-daemon"
version = "0.1.0"
description = "HTTP + WebSocket daemon for Basler cameras"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9",
    "numpy>=1.26",
    "Pillow>=10.0",
    "pypylon>=4.0",
    "pywin32; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev   = ["ruff", "mypy>=1.15", "pytest", "pytest-asyncio>=1.0"]
build = ["pyinstaller"]

[project.scripts]
basler-daemon = "basler_camera_daemon.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/basler_camera_daemon"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["basler_camera_daemon"]

[tool.mypy]
strict = true
python_version = "3.13"
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = "pypylon.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["win32service", "win32serviceutil", "win32event", "win32api", "pywintypes"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
